Added ReadMe file
Creating this and each reading might help us all stay on same page. I've put a brief, unfinished description of the programs functionality, PLEASE ADD MORE as you feel we need. 
This is a Team Project so we should all agree on the path of our application. Your input is very important.